parquet: drop dhat from the page-spill memory test
Replace the experimental, single-maintainer dhat crate with the in-tree
thread-local tracking-allocator pattern already used by
parquet/benches/arrow_reader_peak_memory.rs, and fold the test into the
existing arrow_writer test binary instead of a dedicated one (saving a
compile/link). The measurement still observes real process peak heap and
the assertions are unchanged.
